Round abs small number to 0 and big to infty
Since expn may be summed to negative and it is an unsigned number, it
was considered greater than 0x7FFu.

A normal number has expn in range (0, 1023), so the number with expn
greater than or equal to 1023 is set to \infty, and the number with expn
less than or equal 0 is set to value 0, with one of the signs.

TODO:
Write subnormal (denormalized) format.
Find out why there is difference at MSB of mtsa.